requirements.txt
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
build==1.2.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
distlib==0.3.8
dnspython==2.6.1
filelock==3.15.4
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==5.0.0
future==1.0.0
gevent==24.2.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.62.3
gunicorn==23.0.0
httplib2==0.22.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
load-dotenv==0.1.0
MarkupSafe==2.1.5
mistune==3.0.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
pymongo==4.8.0
pyparsing==3.1.4
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
rpds-py==0.20.0
rsa==4.9
setuptools==74.1.2
six==1.16.0
tldextract==5.1.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.0.3
